# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MaiPai Stack API",
    "version": "0.1.0"
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      }
    },
...
  },
  "paths": {
...
    "/stack/v1/models": {
      "get": {
        "tags": [
          "Models"
        ],
        "summary": "Installed and known models",
        "responses": {
          "200": {
            "description": "Every model record with its provenance and state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "models": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "roles": {
                            "type": "array",
                            "items": {
                              "type": "string"
                            }
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "notInstalled",
                              "installed"
                            ]
                          },
                          "runtimeState": {
                            "type": "string",
                            "enum": [
                              "loaded",
                              "ready"
                            ]
                          },
                          "pinned": {
                            "type": "boolean"
                          },
                          "sizeBytes": {
                            "type": "integer",
                            "nullable": true
                          },
                          "fileMissing": {
                            "type": "boolean"
                          },
                          "measuredFootprintBytes": {
                            "type": "integer",
                            "nullable": true
                          },
                          "measuredContextLength": {
                            "type": "integer",
                            "nullable": true
                          },
                          "estimated": {
                            "type": "boolean"
                          },
                          "source": {
                            "type": "string"
                          },
                          "licence": {
                            "type": "string",
                            "nullable": true
                          },
                          "revision": {
                            "type": "string"
                          },
                          "sha256": {
                            "type": "string",
                            "nullable": true
                          },
                          "provenance": {
                            "type": "object",
                            "additionalProperties": {
                              "nullable": true
                            }
                          },
                          "modelPath": {
                            "type": "string",
                            "nullable": true
                          },
                          "installedAt": {
                            "type": "string",
                            "nullable": true
                          },
                          "verifiedAt": {
                            "type": "string",
                            "nullable": true
                          }
                        },
                        "required": [
                          "id",
                          "roles",
                          "state",
                          "runtimeState",
                          "pinned",
                          "sizeBytes",
                          "fileMissing",
                          "measuredFootprintBytes",
                          "measuredContextLength",
                          "estimated",
                          "source",
                          "licence",
                          "revision",
                          "sha256",
                          "provenance",
                          "modelPath",
                          "installedAt",
                          "verifiedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "models"
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Models"
        ],
        "summary": "Install a pinned model (progress on the event feed)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "chat",
                      "coding",
                      "judge",
                      "router",
                      "embed",
                      "rerank",
                      "vision",
                      "stt",
                      "tts",
                      "wakeword",
                      "image",
                      "video",
                      "music"
                    ]
                  },
                  "repo": {
                    "type": "string"
                  },
                  "url": {
                    "type": "string",
                    "format": "uri"
                  },
                  "sha256": {
                    "type": "string",
                    "minLength": 64,
                    "maxLength": 64
                  },
                  "approx_bytes": {
                    "type": "integer",
                    "minimum": 0
                  },
                  "licence": {
                    "type": "string"
                  },
                  "revision": {
                    "type": "string"
                  },
                  "engine": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "role",
                  "url",
                  "sha256",
                  "licence",
                  "revision"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "The download job.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": {
                      "type": "string"
                    },
                    "model": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "job",
                    "model"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Provenance incomplete.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/stack/v1/jobs": {
      "post": {
        "tags": [
          "Jobs"
        ],
        "summary": "Submit a generator job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "kind": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string",
                    "enum": [
                      "chat",
                      "coding",
                      "judge",
                      "router",
                      "embed",
                      "rerank",
                      "vision",
                      "stt",
                      "tts",
                      "wakeword",
                      "image",
                      "video",
                      "music"
                    ]
                  },
                  "input": {
                    "type": "object",
                    "additionalProperties": {
                      "nullable": true
                    }
                  }
                },
                "required": [
                  "kind"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Queued; progress arrives on the event feed.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "job": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "kind": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string",
                          "nullable": true
                        },
                        "state": {
                          "type": "string",
                          "enum": [
                            "queued",
                            "running",
                            "done",
                            "failed",
                            "cancelled"
                          ]
                        },
                        "percent": {
                          "type": "integer",
                          "minimum": 0,
                          "maximum": 100
                        },
                        "completedBytes": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "totalBytes": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "status": {
                          "type": "string"
                        },
                        "input": {
                          "type": "object",
                          "nullable": true,
                          "additionalProperties": {
                            "nullable": true
                          }
                        },
                        "result": {
                          "nullable": true
                        },
                        "reason": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "kind",
                        "role",
                        "state",
                        "percent",
                        "completedBytes",
                        "totalBytes",
                        "status",
                        "input",
                        "reason",
                        "createdAt",
                        "updatedAt"
                      ]
                    }
                  },
                  "required": [
                    "job"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Nothing on this machine can run the kind.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Recent jobs",
        "responses": {
          "200": {
            "description": "Running and recently finished jobs.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "jobs": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "kind": {
                            "type": "string"
                          },
                          "role": {
                            "type": "string",
                            "nullable": true
                          },
                          "state": {
                            "type": "string",
                            "enum": [
                              "queued",
                              "running",
                              "done",
                              "failed",
                              "cancelled"
                            ]
                          },
                          "percent": {
                            "type": "integer",
                            "minimum": 0,
                            "maximum": 100
                          },
                          "completedBytes": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "totalBytes": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "status": {
                            "type": "string"
                          },
                          "input": {
                            "type": "object",
                            "nullable": true,
                            "additionalProperties": {
                              "nullable": true
                            }
                          },
                          "result": {
                            "nullable": true
                          },
                          "reason": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id",
                          "kind",
                          "role",
                          "state",
                          "percent",
                          "completedBytes",
                          "totalBytes",
                          "status",
                          "input",
                          "reason",
                          "createdAt",
                          "updatedAt"
                        ]
                      }
                    }
                  },
                  "required": [
                    "jobs"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}